Removed the streams watching for order deletions that then triggered a rollback where the order id was passed into the trackingInvemtory model.
The flow is now to create the order document after the payment has been made using stripe api for pre auth, Then Inventory is deducted fromitems.stoclLevel, Then the payment is capture

If the payment is successful a webhook triggers and send a rabbitmq event that send the item.quantity of each item to the product and user services to update the product.stockLevels to match the order services prodcuts.stockLevls.

If the order fails for whatever reason the rollBackInventory function will trigger to restore the products.stockLevels of each item.

created a client pod which will be used to test the stripe flow end to end

Added the connectToRabbitMq function to the server start function so the servers are able to successfully sub to rabbitmq queues straight away once its available